fix: more cleanup of the custom reporter
