+++ package.json
@@ -131,13 +131,13 @@
     "scm:push:gitlab-mirror:commits": "git push gitlab-mirror",
     "scm:push:gitlab-mirror:tags": "git push --tags gitlab-mirror",
     "scm:stage": "git add .",
-    "test": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings node --test --test-reporter ./tools/custom-reporter.mjs test/*.test.mjs test/*/*.test.mjs test/*/*/*.test.mjs test/*/*/*/*.test.mjs",
+    "test": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings node --test --test-reporter ./tools/dots-with-summary.reporter.mjs test/*.test.mjs test/*/*.test.mjs test/*/*/*.test.mjs test/*/*/*/*.test.mjs",
     "test:mocha": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings mocha",
-    "test:r": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings node --test --test-reporter ./tools/custom-reporter.mjs test/*.test.mjs",
-    "test:i": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings node --test --test-name-pattern=\"^\\[[I]\\]\" --test-reporter ./tools/custom-reporter.mjs test/*.test.mjs test/*/*.test.mjs test/*/*/*.test.mjs test/*/*/*/*.test.mjs",
-    "test:u": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings node --test --test-name-pattern=\"^\\[[U]\\]\" --test-reporter ./tools/custom-reporter.mjs test/*.test.mjs test/*/*.test.mjs test/*/*/*.test.mjs test/*/*/*/*.test.mjs",
-    "test:e": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings node --test --test-name-pattern=\"^\\[[E]\\]\" --test-reporter ./tools/custom-reporter.mjs test/*.test.mjs test/*/*.test.mjs test/*/*/*.test.mjs test/*/*/*/*.test.mjs",
-    "test:cover": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings c8 node --test --test-reporter ./tools/custom-reporter.mjs test/*.test.mjs test/*/*.test.mjs test/*/*/*.test.mjs test/*/*/*/*.test.mjs",
+    "test:r": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings node --test --test-reporter ./tools/dots-with-summary.reporter.mjs test/*.test.mjs",
+    "test:i": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings node --test --test-name-pattern=\"^\\[[I]\\]\" --test-reporter ./tools/dots-with-summary.reporter.mjs test/*.test.mjs test/*/*.test.mjs test/*/*/*.test.mjs test/*/*/*/*.test.mjs",
+    "test:u": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings node --test --test-name-pattern=\"^\\[[U]\\]\" --test-reporter ./tools/dots-with-summary.reporter.mjs test/*.test.mjs test/*/*.test.mjs test/*/*/*.test.mjs test/*/*/*/*.test.mjs",
+    "test:e": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings node --test --test-name-pattern=\"^\\[[E]\\]\" --test-reporter ./tools/dots-with-summary.reporter.mjs test/*.test.mjs test/*/*.test.mjs test/*/*/*.test.mjs test/*/*/*/*.test.mjs",
+    "test:cover": "LANG=en_US.UTF-8 NODE_OPTIONS=--no-warnings c8 node --test --test-reporter ./tools/dots-with-summary.reporter.mjs test/*.test.mjs test/*/*.test.mjs test/*/*/*.test.mjs test/*/*/*/*.test.mjs",
     "test:glob": "set -f && test \"`bin/dependency-cruise.mjs --no-config test/extract/__mocks__/gather-globbing/packages/**/src/**/*.js | grep \"no dependency violations found\"`\" = \"✔ no dependency violations found (6 modules, 0 dependencies cruised)\"",
     "test:load": "hyperfine --warmup 3 --runs 30 \"bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress\"",
     "test:load:short": "hyperfine --warmup 1 --runs 5 \"bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress\"",
